requirements.txt
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
scikit-learn==1.3.0
shap==0.42.1
streamlit==1.25.0
xgboost==1.7.6
joblib==1.3.2
pyyaml>=6.0
